Rename mapbox in mapbox-gl
